# swagger.yaml
swagger: "2.0"
info:
  title: Bank Argent API documentation
  description: Contains all available API endpoints in this codebase
  version: "1.0.0"
  termsOfService: "http://swagger.io/terms/"
host: localhost:3001
basePath: /api/v1
schemes:
  - http

paths:
  /user/login:
    post:
      tags:
        - User Module
      summary: Login
      description: API for Login
      parameters:
        - in: body
          name: body
          description: Login Payload
          required: true
          schema:
            $ref: "#/definitions/Login"
      produces:
        - application/json
      responses:
        "200":
          description: Login Successfully
          schema:
            $ref: "#/definitions/LoginResponse"
        "400":
          description: Invalid Fields
        "500":
          description: Internal Server Error

  /user/signup:
    post:
      tags:
        - User Module
      summary: Signup
      description: API for Signup
      parameters:
        - in: body
          name: body
          description: Signup Payload
          required: true
          schema:
            $ref: "#/definitions/User"
      produces:
        - application/json
      responses:
        "200":
          description: Signup Successfully
          schema:
            $ref: "#/definitions/UserApiResponse"
        "400":
          description: Invalid Fields
        "500":
          description: Internal Server Error

  /user/profile:
    post:
      security:
        - Bearer: []
      tags:
        - User Module
      summary: User Profile API
      description: API for fetching a user profile
      #produces:
      # - application/json
      responses:
        "200":
          description: User profile retrieved successully
          schema:
            $ref: "#/definitions/UserResponse"
        "400":
          description: Invalid Fields
        "500":
          description: Internal Server Error
    put:
      security:
        - Bearer: []
      tags:
        - User Module
      summary: User Profile API
      description: API for updating a user profile
      parameters:
        - in: body
          name: body
          description: Update user profile attributes
          required: true
          schema:
            $ref: "#/definitions/UserProfile"
      produces:
        - application/json
      responses:
        "200":
          description: User profile retrieved successully
          schema:
            $ref: "#/definitions/UserResponse"
        "400":
          description: Invalid Fields
        "500":
          description: Internal Server Error

  /user/account:
    put:
      security:
        - Bearer: []
      tags:
        - Admin Module
      summary: User Profile API
      description: API for add account at a user profile
      parameters:
        - in: body
          name: body
          description: account payload
          required: true
          schema:
            $ref: "#/definitions/UserAccount"
      produces:
        - application/json
      responses:
        "200":
          description: add account at a user profile successully
          schema:
            $ref: "#/definitions/accountResponse"
        "400":
          description: Invalid Fields
        "500":
          description: Internal Server Error

  /user/account/operations:
    put:
      security:
        - Bearer: []
      tags:
        - Admin Module
      summary: User Profile API
      description: API for add operation in account at a user profile
      parameters:
        - in: body
          name: body
          description: add operation in account at a user profile
          required: true
          schema:
            $ref: "#/definitions/operation"
      produces:
        - application/json
      responses:
        "200":
          description: add operations at a user profile successully
          schema:
            $ref: "#/definitions/operationResponse"
        "400":
          description: Invalid Fields
        "500":
          description: Internal Server Error

  /user/account/operations/description:
    put:
      security:
        - Bearer: []
      tags:
        - Admin Module
      summary: User Profile API
      description: API for add or update descrition of operations
      parameters:
        - in: body
          name: body
          description: add or update descrition of operations
          required: true
          schema:
            $ref: "#/definitions/description"
      produces:
        - application/json
      responses:
        "200":
          description: add or update descrition successully
          schema:
            $ref: "#/definitions/description"
        "400":
          description: Invalid Fields
        "500":
          description: Internal Server Error

  /user/account/close:
    put:
      security:
        - Bearer: []
      tags:
        - Admin Module
      summary: User Profile API
      description: API for close account at a user profile
      parameters:
        - in: body
          name: body
          description: close account at a user profile
          required: true
          schema:
            $ref: "#/definitions/visible"
      produces:
        - application/json
      responses:
        "200":
          description: close account at a user profile successully
          schema:
            $ref: "#/definitions/visible"
        "400":
          description: Invalid Fields
        "500":
          description: Internal Server Error

securityDefinitions:
  Bearer:
    type: apiKey
    name: Authorization
    in: header

definitions:
  User:
    properties:
      email:
        type: string
        description: user email
      password:
        type: string
        description: user password
      firstName:
        type: string
        description: user first name
      lastName:
        type: string
        description: user last name
      userName:
        type: string
        description: user public name
  Login:
    properties:
      email:
        type: string
        description: user email
      password:
        type: string
        description: user password
  LoginResponse:
    type: object
    properties:
      token:
        type: string
  UserProfile:
    type: object
    properties:
      userName:
        type: string
  UserAccount:
    type: object
    properties:
      name:
        type: string
      nbAccount:
        type: string
      solde:
        type: number
      visible:
        type: boolean
      operations:
        type: array
        items:
          type: object
  operation:
    type: object
    properties:
      date:
        type: string
      title:
        type: string
      description:
        type: string
      montant:
        type: number
  description:
    type: object
    properties:
      description:
        type: string
  visible:
    type: object
    properties:
      visible:
        type: boolean

  UserResponse:
    type: object
    properties:
      status:
        type: integer
      message:
        type: string
      body:
        type: object
        properties:
          id:
            type: string
          email:
            type: string
  accountResponse:
    type: object
    properties:
      status:
        type: integer
      message:
        type: string
      body:
        type: object
        properties:
          id:
            type: string
          name:
            type: string
          nbAccount:
            type: string
          solde:
            type: number
          visible:
            type: boolean
          operations:
            type: array
            items:
              type: object
  operationResponse:
    type: object
    properties:
      status:
        type: integer
      message:
        type: string
      body:
        type: object
        properties:
          id:
            type: string
          date:
            type: string
          title:
            type: string
          description:
            type: string
          montant:
            type: number